requirements.txt
discord.py>=2.3.0
asyncpg>=0.29.0
feedparser>=6.0.10
aiohttp>=3.9.0
python-dotenv>=1.0.0
